chore: remove obsolete helpers module
