+++ src/tnfr/dynamics/sampling.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from ..helpers import _cache_node_list
+from ..helpers.cache import _cache_node_list
 from ..rng import _rng_for_step, base_seed
 
 __all__ = ["update_node_sample"]
