+++ src/tnfr/gamma.py
@@ -12,7 +12,7 @@ from collections.abc import Mapping
 
 from .constants import ALIAS_THETA
 from .alias import get_attr
-from .helpers import node_set_checksum, edge_version_cache, get_graph_mapping
+from .helpers.cache import node_set_checksum, edge_version_cache, get_graph_mapping
 from .logging_utils import get_logger
