+++ tests/test_coherence_cache.py
@@ -3,7 +3,7 @@ import networkx as nx
 
 from tnfr.constants import THETA_PRIMARY
 from tnfr.metrics import coherence_matrix, local_phase_sync_weighted
-from tnfr.helpers import ensure_node_index_map
+from tnfr.helpers.cache import ensure_node_index_map
 
 
 def make_graph(offset=0):
