+++ tests/test_edge_version_cache_disable.py
@@ -1,6 +1,6 @@
 import networkx as nx
 
-from tnfr.helpers import edge_version_cache
+from tnfr.helpers.cache import edge_version_cache
 
 
 def test_edge_version_cache_disable():
