+++ tests/test_edge_version_cache_reentrant.py
@@ -1,5 +1,5 @@
 import networkx as nx
-from tnfr.helpers import edge_version_cache
+from tnfr.helpers.cache import edge_version_cache
 
 
 def test_edge_version_cache_reentrant():
