+++ tests/test_edge_version_cache_threadsafe.py
@@ -1,7 +1,7 @@
 import networkx as nx
 from concurrent.futures import ThreadPoolExecutor
 
-from tnfr.helpers import edge_version_cache, increment_edge_version
+from tnfr.helpers.cache import edge_version_cache, increment_edge_version
 
 
 def test_edge_version_cache_thread_safety():
