+++ tests/test_neighbors_map_cache.py
@@ -2,7 +2,7 @@ import networkx as nx
 from types import MappingProxyType
 
 from tnfr.metrics_utils import ensure_neighbors_map
-from tnfr.helpers import increment_edge_version
+from tnfr.helpers.cache import increment_edge_version
 
 
 def test_neighbors_map_reuses_proxy():
