+++ tests/test_node_set_checksum.py
@@ -2,8 +2,13 @@ import hashlib
 import networkx as nx
 import timeit
 
-from tnfr import helpers as h
-from tnfr.helpers import node_set_checksum, _stable_json, increment_edge_version
+from tnfr.helpers.cache import (
+    node_set_checksum,
+    _stable_json,
+    increment_edge_version,
+    _node_repr,
+    _hash_node,
+)
 
 
 def build_graph():
@@ -30,7 +35,7 @@ def test_node_set_checksum_object_stable():
 
 
 def _reference_checksum(G):
-    nodes = sorted(G.nodes(), key=h._node_repr)
+    nodes = sorted(G.nodes(), key=_node_repr)
     hasher = hashlib.blake2b(digest_size=16)
     for n in nodes:
         d = hashlib.blake2b(
@@ -57,7 +62,7 @@ def test_node_set_checksum_presorted_performance():
     G = nx.Graph()
     G.add_nodes_from(range(1000))
     nodes = list(G.nodes())
-    nodes.sort(key=h._node_repr)
+    nodes.sort(key=_node_repr)
     t_unsorted = timeit.timeit(lambda: node_set_checksum(G, nodes), number=1)
     t_presorted = timeit.timeit(
         lambda: node_set_checksum(G, nodes, presorted=True), number=1
@@ -74,23 +79,23 @@ def test_node_set_checksum_no_store_does_not_cache():
 
 def test_node_repr_cache_cleared_on_increment():
     nxG = nx.Graph()
-    h._node_repr("foo")
-    assert h._node_repr.cache_info().currsize > 0
+    _node_repr("foo")
+    assert _node_repr.cache_info().currsize > 0
     increment_edge_version(nxG)
-    assert h._node_repr.cache_info().currsize == 0
+    assert _node_repr.cache_info().currsize == 0
 
 
 def test_hash_node_cache_cleared_on_increment():
     nxG = nx.Graph()
-    h._hash_node(("foo", 1))
-    assert h._hash_node.cache_info().currsize > 0
+    _hash_node(("foo", 1))
+    assert _hash_node.cache_info().currsize > 0
     increment_edge_version(nxG)
-    assert h._hash_node.cache_info().currsize == 0
+    assert _hash_node.cache_info().currsize == 0
 
 
 def test_hash_node_matches_manual():
     obj = ("a", 1)
     manual = hashlib.blake2b(
-        h._node_repr(obj).encode("utf-8"), digest_size=16
+        _node_repr(obj).encode("utf-8"), digest_size=16
     ).digest()
-    assert h._hash_node(obj) == manual
+    assert _hash_node(obj) == manual
