+++ tests/test_trig_cache_reuse.py
@@ -4,7 +4,7 @@ import pytest
 
 from tnfr.constants import ALIAS_THETA, ALIAS_VF, ALIAS_DNFR
 from tnfr.metrics_utils import get_trig_cache, compute_Si
-from tnfr.helpers import neighbor_phase_mean
+from tnfr.helpers.numeric import neighbor_phase_mean
 from tnfr.alias import set_attr
